+++ packages/gatsby-plugin-image/src/components/gatsby-image.server.tsx
@@ -1,5 +1,5 @@
 import React, { ElementType, FunctionComponent, CSSProperties } from "react"
-import { GatsbyImageProps } from "./gatsby-image.browser"
+import { GatsbyImageProps, ISharpGatsbyImageData } from "./gatsby-image.browser"
 import { getWrapperProps, getMainProps, getPlaceHolderProps } from "./hooks"
 import { Placeholder } from "./placeholder"
 import { MainImage, MainImageProps } from "./main-image"
@@ -19,21 +19,28 @@ export const GatsbyImage: FunctionComponent<GatsbyImageProps> = function GatsbyI
   as,
   className,
   style,
-  placeholder,
-  images,
-  width,
-  height,
-  layout = `fixed`,
+  image,
   loading = `lazy`,
   ...props
 }) {
+  const {
+    width,
+    height,
+    layout,
+    images,
+    placeholder,
+    sizes,
+    backgroundColor,
+  } = image
+
   const { style: wStyle, className: wClass, ...wrapperProps } = getWrapperProps(
     width,
     height,
-    layout
+    layout,
+    backgroundColor
   )
 
-  const cleanedImages: GatsbyImageProps["images"] = {
+  const cleanedImages: ISharpGatsbyImageData["images"] = {
     fallback: undefined,
     sources: [],
   }
@@ -69,6 +76,7 @@ export const GatsbyImage: FunctionComponent<GatsbyImageProps> = function GatsbyI
         {placeholder && <Placeholder {...getPlaceHolderProps(placeholder)} />}
         <MainImage
           data-gatsby-image-ssr=""
+          sizes={sizes}
           {...(props as Omit<MainImageProps, "images" | "fallback">)}
           // When eager is set we want to start the isLoading state on true (we want to load the img without react)
           {...getMainProps(loading === `eager`, false, cleanedImages, loading)}