+++ packages/gatsby-plugin-image/src/components/static-image.server.tsx
@@ -1,11 +1,11 @@
 import React, { FunctionComponent } from "react"
 import { StaticImageProps } from "../utils"
 import { GatsbyImage as GatsbyImageServer } from "./gatsby-image.server"
-import { GatsbyImageProps } from "./gatsby-image.browser"
+import { GatsbyImageProps, ISharpGatsbyImageData } from "./gatsby-image.browser"
 
 // These values are added by Babel. Do not add them manually
 interface IPrivateProps {
-  __imageData?: GatsbyImageProps
+  __imageData?: ISharpGatsbyImageData
   __error?: string
 }
 
@@ -23,12 +23,12 @@ export function _getStaticImage(
     }
 
     if (imageData) {
-      return <GatsbyImage {...imageData} {...props} />
+      return <GatsbyImage image={imageData} {...props} />
     }
     console.warn(`Image not loaded`, src)
     if (!__error && process.env.NODE_ENV === `development`) {
       console.warn(
-        `Please ensure that "gatsby-plugin-image" is included in the plugins array in gatsby-config.js`
+        `Please ensure that "gatsby-plugin-image" is included in the plugins array in gatsby-config.js, and that your version of gatsby is at least 2.24.78`
       )
     }
     return null