+++ packages/gatsby-plugin-image/src/components/static-image.tsx
@@ -1,12 +1,12 @@
 import {
   GatsbyImage as GatsbyImageBrowser,
-  GatsbyImageProps,
+  ISharpGatsbyImageData,
 } from "./gatsby-image.browser"
 import { _getStaticImage } from "./static-image.server"
 import { StaticImageProps } from "../utils"
 // These values are added by Babel. Do not add them manually
 interface IPrivateProps {
-  __imageData?: GatsbyImageProps
+  __imageData?: ISharpGatsbyImageData
   __error?: string
 }