+++ packages/gatsby-transformer-sharp/src/gatsby-node.js
@@ -1,3 +1,8 @@
-exports.onCreateNode = require(`./on-node-create`)
+const {
+  onCreateNode,
+  unstable_shouldOnCreateNode,
+} = require(`./on-node-create`)
+exports.onCreateNode = onCreateNode
+exports.unstable_shouldOnCreateNode = unstable_shouldOnCreateNode
 exports.createSchemaCustomization = require(`./customize-schema`)
 exports.createResolvers = require(`./create-resolvers`)