+++ packages/gatsby/scripts/__tests__/api.js
@@ -59,6 +59,9 @@ it("generates the expected api output", done => {
           "resolvableExtensions": Object {},
           "setFieldsOnGraphQLNodeType": Object {},
           "sourceNodes": Object {},
+          "unstable_shouldOnCreateNode": Object {
+            "version": "2.24.80",
+          },
         },
         "ssr": Object {
           "onPreRenderHTML": Object {},