+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.43](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.42...gatsby-plugin-sharp@2.6.43) (2020-10-19)
+
+### Performance Improvements
+
+- **gatsby-plugin-image:** Optimise image size functions ([#27544](https://github.com/gatsbyjs/gatsby/issues/27544)) ([8752477](https://github.com/gatsbyjs/gatsby/commit/87524778996452b2649e29588526a724f9813d45))
+
 ## [2.6.42](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.41...gatsby-plugin-sharp@2.6.42) (2020-10-12)
 
 ### Features