+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.21...gatsby-plugin-typescript@2.4.22) (2020-10-15)
+
+### Features
+
+- **gatsby-plugin-typescript:** Add schema ([#27361](https://github.com/gatsbyjs/gatsby/issues/27361)) ([072ef6c](https://github.com/gatsbyjs/gatsby/commit/072ef6c992168c89505143670013d5d8147f4ab4))
+
 ## [2.4.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.20...gatsby-plugin-typescript@2.4.21) (2020-09-28)
 
 **Note:** Version bump only for package gatsby-plugin-typescript