+++ packages/gatsby-remark-images/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.36](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.3.35...gatsby-remark-images@3.3.36) (2020-10-19)
+
+**Note:** Version bump only for package gatsby-remark-images
+
+## [3.3.35](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.3.34...gatsby-remark-images@3.3.35) (2020-10-16)
+
+**Note:** Version bump only for package gatsby-remark-images
+
+## [3.3.34](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.3.33...gatsby-remark-images@3.3.34) (2020-10-15)
+
+**Note:** Version bump only for package gatsby-remark-images
+
 ## [3.3.33](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.3.32...gatsby-remark-images@3.3.33) (2020-10-06)
 
 ### Bug Fixes